Testes relacionados a desativar a conta de um usuário e regras de negocio a isso adicionadas

+++ src/main/java/com/project/clinicaapi/controller/WorkDayController.java
@@ -24,7 +24,7 @@ public class WorkDayController implements WorkDaySwagger {
 
     @Override
     @GetMapping(value = "/{id}")
-    public ResponseEntity<WorkDayResponseDTO> findById(@PathVariable(value = "id") Long id) {
+    public ResponseEntity<WorkDayResponseDTO> findById(@PathVariable(value = "id") Integer id) {
         return ResponseEntity.ok().body(workDayService.findById(id));
     }
