+++ src/main/java/com/project/clinicaapi/dto/request/register/DentistRegisterDTO.java
@@ -27,6 +27,6 @@ public class DentistRegisterDTO extends UserRegisterDTO {
 
     @JsonProperty(value = "workdays")
     @Size(max = 6)
-    private Set<Long> workDays;
+    private Set<Integer> workDays;
 
 }
